organised addText api flow in client & server
