+++ server/controllers/api/textController.js
@@ -2,45 +2,70 @@ const Text = require("../../models/text");
 const Card = require("../../models/card")
 
 const getUserTexts = async (req, res) => {
-    try {
-      const userId = req.params.userId;
-  
-      // texts for the specific user
-      const texts = await Text.find({ user: userId })
-        // .populate("cards") // TODO: controller for cards not set up yet
-        // .populate("tags") // TODO: controller for tags not set up yet
-        .sort({ createdAt: -1 }); // Newest texts first
-  
-      res.status(200).json(texts);
-    } catch (error) {
-      res.status(500).json({ message: "Error fetching texts", error: error.message });
-    }
-  };
+  try {
+    const userId = req.params.userId;
 
-  const deleteUserText = async (req, res) => {
-    const { userId, textId } = req.params;
-  
-    try {
-      // TODO: can't currently test, once card functionality is up, need to see if cards are actually deleted
-      await Card.deleteMany({ text: textId });
-  
-      // This works
-      const text = await Text.findOneAndDelete({
-        _id: textId,
-        user: userId,
-      });
-  
-      if (!text) {
-        return res.status(404).json({ message: "Text not found" });
-      }
-  
-      return res.status(200).json({ message: "Text and associated cards deleted successfully" });
-    } catch (error) {
-      return res.status(500).json({ message: "Server error", error: error.message });
-    }
-  };
+    // texts for the specific user
+    const texts = await Text.find({ user: userId })
+      // .populate("cards") // TODO: controller for cards not set up yet
+      // .populate("tags") // TODO: controller for tags not set up yet
+      .sort({ createdAt: -1 }); // Newest texts first
 
-  module.exports = {
-    getUserTexts,
-    deleteUserText
+    res.status(200).json(texts);
+  } catch (error) {
+    res.status(500).json({ message: "Error fetching texts", error: error.message });
+  }
+};
+
+const deleteUserText = async (req, res) => {
+  const { userId, textId } = req.params;
+
+  try {
+    // TODO: can't currently test, once card functionality is up, need to see if cards are actually deleted
+    await Card.deleteMany({ text: textId });
+
+    // This works
+    const text = await Text.findOneAndDelete({
+      _id: textId,
+      user: userId,
+    });
+
+    if (!text) {
+      return res.status(404).json({ message: "Text not found" });
+    }
+
+    return res.status(200).json({ message: "Text and associated cards deleted successfully" });
+  } catch (error) {
+    return res.status(500).json({ message: "Server error", error: error.message });
+  }
+};
+
+async function addText(req, res) {
+  const { content, title, source } = req.body;
+  const userId = req.user._id;
+
+  console.log('User ID in addText:', userId);
+  console.log('Adding text:', content);
+
+  try {
+    const newText = new Text({
+      user: userId,
+      title,
+      source,
+      content
+    });
+    console.log('New text:', newText);
+    await newText.save();
+    res.status(201).json({ message: 'Text added successfully', text: newText });
+  } catch (error) {
+    console.error('Error saving text:', error);
+    res.status(500).json({ error: 'Failed to add text' });
+  }
+}
+
+
+module.exports = {
+  getUserTexts,
+  deleteUserText,
+  addText
 };
