+++ server/routes/api/textRoutes.js
@@ -1,8 +1,10 @@
 const express = require("express");
-const { getUserTexts, deleteUserText } = require("../../controllers/api/textController");
+const { addText, getUserTexts, deleteUserText } = require("../../controllers/api/textController");
+const { verifyJWT } = require("./../../utils/jwt");
 const router = express.Router();
 
 router.get("/:userId", getUserTexts);
 router.delete("/:userId/text/:textId", deleteUserText);
+router.post('/add', verifyJWT, addText);
 
 module.exports = router;
